using index.ts instead of app.ts
