+++ express/bin/www.ts
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-var app = require('../app');
+var app = require('../index');
 var debug = require('debug')('express:server');
 var http = require('http');
